Revert "refactor: centralize log configuration"
